+++ src/hops/MarkovChain/Proposal/CSmMALAProposal.hpp
@@ -27,7 +27,7 @@ namespace hops {
     class CSmMALAProposal : public Proposal, public ModelType {
     public:
         /**
-         * @brief Constructs proposal mechanism on polytope defined as Ax<b.
+         * @brief Constructs m_proposal mechanism on polytope defined as Ax<b.
          * @param A
          * @param b
          * @param currentState
@@ -149,7 +149,7 @@ namespace hops {
         proposal = state;
 
         if (!ModelType::getDimensionNames().empty()) {
-            assert(ModelType::getDimensionNames().size() == this->state.rows());
+            assert(ModelType::getDimensionNames().size() == static_cast<std::size_t>(this->state.rows()));
             this->dimensionNames = ModelType::getDimensionNames();
         } else {
             this->dimensionNames = hops::createDefaultDimensionNames(this->state.rows());
